+++ src/App.js
@@ -1,13 +1,11 @@
 import React from 'react';
 import {BrowserRouter, Routes, Route} from "react-router-dom";
 import "./App.css";
-
 import Home from "../src/pages/Home";
-import Subject from './pages/subject';
+import Subject from './pages/Subject.js';
 import Speak from './pages/Speak';
 import Feedback from './pages/Feedback';
-import Loading1 from './pages/Loading1';
-import Loading2 from "./pages/Loding2";
+
 
 function App() {
   return (
@@ -18,9 +16,11 @@ function App() {
             <Route path = "/Subject" element={<Subject/>}/>
             <Route path = "/speak" element={<Speak/>}/>
             <Route path = "/Feedback" element = {<Feedback/>}/>
-            <Route path = "/Loading1" element={< Loading1/>}/>
-            <Route path = "/Loading2" element={<Loading2/>}/>
+            
+            
           </Routes>
+          
+          <Subject/>
         </div>
     
     </BrowserRouter>
