+++ src/components/Common/Button.js
@@ -1,3 +1,12 @@
 import React from "react";
 import { Component } from "react";
 import styled from 'styled-components';
+
+const Button =() =>{
+    return (
+        <button />
+    )
+
+}
+
+export default Button;
