+++ src/components/Common/Navbar.js
@@ -4,9 +4,9 @@ import { Component } from "react";
 const Navbar = () =>{
 
     return (
-        <Narbar>
+        
             <img src="..." alt = "..."/>
-        </Narbar>
+        
 
     );
