+++ src/utils/Palette.js
@@ -1,8 +1,8 @@
-const ColorCode = {
+export const ColorCode = {
     MainBlue : "#00558D",
     SelectBlue : "#42A1E0",
     UnSelectBlue : "#ABDDFF",
 
 };
 
-export {ColorCode};
+export default ColorCode;
